Add clickable moderator action buttons with hotkeys and tooltips to sidebar
