+++ packages/lesswrong/components/sunshineDashboard/supermod/ModerationActionButtons.tsx
@@ -1,0 +1,99 @@
+import React from 'react';
+import { defineStyles, useStyles } from '@/components/hooks/useStyles';
+import LWTooltip from '@/components/common/LWTooltip';
+import KeystrokeDisplay from './KeystrokeDisplay';
+import { useModerationUserActions } from './useModerationUserActions';
+
+const styles = defineStyles('ModerationActionButtons', (theme: ThemeType) => ({
+  actionsColumn: {
+    display: 'flex',
+    flexDirection: 'column',
+    marginTop: 8,
+    marginBottom: 12,
+  },
+  actionTooltip: {
+    display: 'block',
+  },
+  actionRow: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    padding: '2px 4px',
+    fontSize: 13,
+    cursor: 'pointer',
+    '&:hover': {
+      backgroundColor: theme.palette.grey[100],
+    },
+  },
+}));
+
+const ModerationActionButtons = ({user, currentUser, addToUndoQueue}: {
+  user: SunshineUsersList;
+  currentUser: UsersCurrent;
+  addToUndoQueue: (actionLabel: string, executeAction: () => Promise<void>) => void;
+}) => {
+  const classes = useStyles(styles);
+
+  const {
+    handleReview,
+    handleApproveCurrentOnly,
+    handleSnooze,
+    handleSnoozeCustom,
+    handleRejectContentAndRemove,
+    handleRestrictAndNotify,
+  } = useModerationUserActions({ selectedUser: user, currentUser, addToUndoQueue });
+
+  const moderatorActionButtons = [
+    {
+      label: 'Approve',
+      keystroke: 'A',
+      tooltip: "Approve this user and all their content. Marks them as reviewed by you, so their future posts and comments no longer need review. Clears any snooze and any flag, and signs an 'Approved' note in their moderator notes.",
+      onClick: handleReview,
+    },
+    {
+      label: 'Snooze 10',
+      keystroke: 'S',
+      tooltip: "Remove this user from the review queue for now; they'll return for review after their next 10 posts or comments. Signs a 'Snooze 10' note in their moderator notes.",
+      onClick: () => handleSnooze(10),
+    },
+    {
+      label: 'Snooze Custom',
+      keystroke: 'Shift+S',
+      tooltip: 'Same as Snooze 10, but opens a dialog to choose how many more posts or comments the user can make before they return to the review queue.',
+      onClick: handleSnoozeCustom,
+    },
+    {
+      label: 'Approve Current Only',
+      keystroke: 'Shift+A',
+      tooltip: "Approve this user's existing unreviewed posts and comments and remove them from the queue, without marking the user as reviewed, so their future content will still need review. Also clears any snooze and any flag.",
+      onClick: handleApproveCurrentOnly,
+    },
+    {
+      label: 'Reject Latest and Remove',
+      keystroke: 'X',
+      tooltip: "Opens the rejection-reason dialog, then rejects this user's most recent unapproved post or comment with that reason, and removes the user from the review queue (without approving them).",
+      onClick: handleRejectContentAndRemove,
+    },
+    {
+      label: 'Reject, Restrict, and Notify',
+      keystroke: 'Shift+R',
+      tooltip: "Opens the rejection-reason dialog, then a message dialog. Rejects this user's most recent unapproved post or comment, disables their posting, commenting, messaging, and voting, sends them the message as a moderator PM, and removes them from the review queue.",
+      onClick: handleRestrictAndNotify,
+    },
+  ];
+
+  return (
+    <div className={classes.actionsColumn}>
+      {moderatorActionButtons.map(({label, keystroke, tooltip, onClick}) => (
+        <LWTooltip key={label} title={tooltip} placement="left" className={classes.actionTooltip}>
+          <div className={classes.actionRow} onClick={onClick}>
+            <span>{label}</span>
+            <KeystrokeDisplay keystroke={keystroke} splitBeforeTranslation />
+          </div>
+        </LWTooltip>
+      ))}
+    </div>
+  );
+};
+
+export default ModerationActionButtons;
